pyproject.toml
[project]
name = "battlelens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.100.2",
    "openai-agents>=0.2.8",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.9",
    "tavily-python>=0.7.11",
]
